wedlock_app/requirements.txt
appdirs==1.4.0
backcall==0.1.0
bleach==3.0.2
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.3.7
coverage==4.3.4
cycler==0.10.0
dataclasses==0.7
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.40.0
django-cors-headers==3.1.0
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-tastypie==0.14.2
djangorestframework==3.10.3
djongo==1.2.36
dnspython==1.16.0
entrypoints==0.2.3
Flask==1.0.2
gunicorn==20.0.4
httpie==1.0.3
idna==2.8
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==0.23
matplotlib==3.0.0
missingno==0.4.1
mistune==0.8.4
mongoengine==0.8.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.3
oauthlib==3.1.0
packaging==16.8
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
Pillow==6.1.0
pipenv==2018.11.26
prometheus-client==0.4.2
prompt-toolkit==2.0.6
psycopg2==2.8.5
py==1.4.32
pyflakes==1.5.0
Pygments==2.4.2
pymongo==3.9.0
pyparsing==2.1.10
pytest==3.0.6
python-dateutil==2.6.0
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2018.6
pywinpty==0.5.4
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.21.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sklearn==0.0
sqlparse==0.2.4
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.25.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==5.0.1
widgetsnbextension==3.4.2
